Updates the computation of the distance/proximity and probabilistic selection
Previously, the computation wasn't totally right. In fact, the dot product was
made on non-normalized vectors. Thus, the result can be great than 1 or less
than -1. The case of a negative result wasn't also treated.
Now, the dot product result is normalized. And to cast the result in [0,1],
the mean is taken from the result added to 1.

The probability assigned to a track is now its proximity (in [0,1]), instead of
1 - distance.